Settings: Slims navbar and navring customizations
add back our navbar and navring tweeks. This is a completly overworked
version compared to the one we carry about a year now. Small glitches
we had are now all fixed. As well it does not need to register extra
observers
code was much slimmed down and improved.
Beside that we adapt the 4.4 navbar changes to our system so that it it
consistent with the new 4.4 UI
This patch adds:
- Navbar up to 5 buttons
- navring up to 5 buttons
- dynamic menu behaviour
- landscape navbar options as well now on tablets
- button glow color
- button icon color
- navrign and navbar user can set system or custom icons
- navbar height and width
- enable/disable navbar and navring
- actions are vor navbar and available for the whole system over
SlimActions.java
back
custom app or activtiy
home
kill app
last app
notification drawer
ime switcher
power menu
qs pannel
recents screen
screen off
screenshot
search
torch if it is available on the device
- navring has targets for handling vib/mute/sound
- added Color class based on blk_jacks idea which first grey an icon out
and then color it. This gives way better results. As well it fits now
perfectly with the new 4.4 UI. Put it on a central place so that it is
from everywhere available. notification shortcuts adapted to the new
class
place
- complete new icons thanks to kroz who has redone all for xxhdpi
- our slim special swipe drag drop press longpress options support
like on 4.3 this two features
- a lot lot fixes as well on current google code